implement of q2 and q4
Write the q2 with binary search. If find the character in the array,
return the first character larger than the input character. If not
found, return the smallest character that is strictly larger than the
search character. If no such character exists, return the smallest
character in the array.

Write the q4 with dynamic programming. Find the contiguous subarray
within an array which has the largest sum.